merge LoggerBackendProvider with LoggerFactory
LoggerFactory was a simple facade for LoggerBackendProvider
which could be implemented in the only way.
Also it could not be decorated because it returns Logger.

Now LoggerFactory is an abstract class with single abstract
method from LoggerBackendProvider (which is removed).
It is composable since this method returns LoggerBackend
interface which can be decoraded in any way.

DefaultLoggerFactory is public now and can be used in
custom factories. The factory itself just delegates to
Slf4jBackend.Factory.