big changes in LoggerContext: MessageDecor
Introduced MessageDecor: an object that is responsible for
decorating messages with context data.
All loggers obtained from a LoggerFactory will use
MessageDecor provided by that factory.
LoggingContext contains a MessageDecor bound to a factory
that provided it for performance considerations,
however different decors can be used in the same context.