transform Logger to CoroutineLogger
Using a logger with an empty context as a default seems to be wrong.
For this reason now former Logger class is represented by
ContextLogger (to indicate that it has a context attached to it)
and former CoroutineLogger class is represented by Logger itself.
Thus logger uses coroutine context by default and
must be explicitly transformed to a logger with attached empty context.

Changed version to 0.10 because of this huge breaking change.