add MessageBuilder for future extensions
Future versions of the library may allow users to specify
additional data inside the lambda passed to log { } method.

For example:

    logger.info { marker(ALERT); "An event occured" }

This change brings an incompatibility:
'this' may not be used there anymore the way it worked in previous versions,
as it always points to a MessageBuilder object now.
Example of a broken code:

    logger.info { "$this is fine" }